Use the correct typing_extensions version
The extra pinning in the run section comes from upstream. The switch from
typing-extensions to typing_extensions is because of awaiting on a current fix
for typing-extensions 4.9.0 (see diff between build 0 and build 1 in that
package for more details). Underscore works fine; may not revert back to dash
because no need.